Move get_model_arch to util.py
